Modularize reconstruction I/O formats into different libraries
