+++ src/colmap/scene/reconstruction_io.cc
@@ -29,13 +29,6 @@
 
 #include "colmap/scene/reconstruction_io.h"
 
-#include "colmap/geometry/rigid3.h"
-#include "colmap/scene/camera.h"
-#include "colmap/scene/image.h"
-#include "colmap/scene/point2d.h"
-#include "colmap/scene/point3d.h"
-#include "colmap/scene/reconstruction.h"
-#include "colmap/scene/track.h"
 #include "colmap/util/file.h"
 #include "colmap/util/misc.h"
 #include "colmap/util/ply.h"
@@ -44,1178 +37,6 @@
 #include <fstream>
 
 namespace colmap {
-namespace {
-
-// Helper method to extract sorted camera, image, point3D identifiers.
-// We sort the identifiers before writing to the stream, such that we produce
-// deterministic output independent of standard library dependent ordering of
-// the unordered map container.
-template <typename ID_TYPE, typename DATA_TYPE>
-std::vector<ID_TYPE> ExtractSortedIds(
-    const std::unordered_map<ID_TYPE, DATA_TYPE>& data,
-    const std::function<bool(const DATA_TYPE&)>& filter = nullptr) {
-  std::vector<ID_TYPE> ids;
-  ids.reserve(data.size());
-  for (const auto& [id, d] : data) {
-    if (filter == nullptr || filter(d)) {
-      ids.push_back(id);
-    }
-  }
-  std::sort(ids.begin(), ids.end());
-  return ids;
-}
-
-void CreateOneRigPerCamera(Reconstruction& reconstruction) {
-  for (const auto& [camera_id, camera] : reconstruction.Cameras()) {
-    Rig rig;
-    rig.SetRigId(camera_id);
-    rig.AddRefSensor(camera.SensorId());
-    reconstruction.AddRig(std::move(rig));
-  }
-}
-
-void CreateFrameForImage(const Image& image,
-                         const Rigid3d& cam_from_world,
-                         Reconstruction& reconstruction) {
-  Frame frame;
-  frame.SetFrameId(image.ImageId());
-  frame.SetRigId(image.CameraId());
-  frame.AddDataId(image.DataId());
-  frame.SetRigFromWorld(cam_from_world);
-  reconstruction.AddFrame(std::move(frame));
-}
-
-std::unordered_map<image_t, Frame*> ExtractImageToFramePtr(
-    Reconstruction& reconstruction) {
-  std::unordered_map<image_t, Frame*> image_to_frame;
-  for (const auto& [frame_id, frame] : reconstruction.Frames()) {
-    for (const data_t& data_id : frame.DataIds()) {
-      if (data_id.sensor_id.type == SensorType::CAMERA) {
-        THROW_CHECK(
-            image_to_frame.emplace(data_id.id, &reconstruction.Frame(frame_id))
-                .second);
-      }
-    }
-  }
-  return image_to_frame;
-}
-
-}  // namespace
-
-void ReadRigsText(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  std::string line;
-  std::string item;
-
-  while (std::getline(stream, line)) {
-    StringTrim(&line);
-
-    if (line.empty() || line[0] == '#') {
-      continue;
-    }
-
-    std::stringstream line_stream(line);
-
-    Rig rig;
-
-    // ID
-    std::getline(line_stream, item, ' ');
-    rig.SetRigId(std::stoull(item));
-
-    // NUM_SENSORS
-    std::getline(line_stream, item, ' ');
-    const uint64_t num_sensors = std::stoull(item);
-
-    if (num_sensors > 0) {
-      // REF_SENSOR
-      sensor_t ref_sensor_id;
-      std::getline(line_stream, item, ' ');
-      ref_sensor_id.type = SensorTypeFromString(item);
-      std::getline(line_stream, item, ' ');
-      ref_sensor_id.id = std::stoul(item);
-      rig.AddRefSensor(ref_sensor_id);
-    }
-
-    // SENSORS
-    if (num_sensors > 1) {
-      for (uint64_t i = 0; i < num_sensors - 1; ++i) {
-        sensor_t sensor_id;
-        std::getline(line_stream, item, ' ');
-        sensor_id.type = SensorTypeFromString(item);
-        std::getline(line_stream, item, ' ');
-        sensor_id.id = std::stoul(item);
-
-        std::getline(line_stream, item, ' ');
-        const bool has_pose = item == "1";
-
-        std::optional<Rigid3d> sensor_from_rig;
-        if (has_pose) {
-          sensor_from_rig = Rigid3d();
-
-          std::getline(line_stream, item, ' ');
-          sensor_from_rig->rotation.w() = std::stold(item);
-
-          std::getline(line_stream, item, ' ');
-          sensor_from_rig->rotation.x() = std::stold(item);
-
-          std::getline(line_stream, item, ' ');
-          sensor_from_rig->rotation.y() = std::stold(item);
-
-          std::getline(line_stream, item, ' ');
-          sensor_from_rig->rotation.z() = std::stold(item);
-
-          std::getline(line_stream, item, ' ');
-          sensor_from_rig->translation.x() = std::stold(item);
-
-          std::getline(line_stream, item, ' ');
-          sensor_from_rig->translation.y() = std::stold(item);
-
-          std::getline(line_stream, item, ' ');
-          sensor_from_rig->translation.z() = std::stold(item);
-        }
-
-        rig.AddSensor(sensor_id, sensor_from_rig);
-      }
-    }
-
-    reconstruction.AddRig(std::move(rig));
-  }
-}
-
-void ReadRigsText(Reconstruction& reconstruction, const std::string& path) {
-  std::ifstream file(path);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadRigsText(reconstruction, file);
-}
-
-void ReadCamerasText(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  std::string line;
-  std::string item;
-
-  while (std::getline(stream, line)) {
-    StringTrim(&line);
-
-    if (line.empty() || line[0] == '#') {
-      continue;
-    }
-
-    std::stringstream line_stream(line);
-
-    struct Camera camera;
-
-    // ID
-    std::getline(line_stream, item, ' ');
-    camera.camera_id = std::stoul(item);
-
-    // MODEL
-    std::getline(line_stream, item, ' ');
-    camera.model_id = CameraModelNameToId(item);
-
-    // WIDTH
-    std::getline(line_stream, item, ' ');
-    camera.width = std::stoll(item);
-
-    // HEIGHT
-    std::getline(line_stream, item, ' ');
-    camera.height = std::stoll(item);
-
-    // PARAMS
-    camera.params.reserve(CameraModelNumParams(camera.model_id));
-    while (!line_stream.eof()) {
-      std::getline(line_stream, item, ' ');
-      camera.params.push_back(std::stold(item));
-    }
-
-    THROW_CHECK(camera.VerifyParams());
-    reconstruction.AddCamera(std::move(camera));
-  }
-}
-
-void ReadCamerasText(Reconstruction& reconstruction, const std::string& path) {
-  std::ifstream file(path);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadCamerasText(reconstruction, file);
-}
-
-void ReadFramesText(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  std::string line;
-  std::string item;
-
-  while (std::getline(stream, line)) {
-    StringTrim(&line);
-
-    if (line.empty() || line[0] == '#') {
-      continue;
-    }
-
-    std::stringstream line_stream(line);
-
-    Frame frame;
-
-    // ID
-    std::getline(line_stream, item, ' ');
-    frame.SetFrameId(std::stoul(item));
-
-    // RIG_ID
-    std::getline(line_stream, item, ' ');
-    frame.SetRigId(std::stoul(item));
-
-    // RIG_FROM_WORLD
-
-    Rigid3d rig_from_world;
-
-    std::getline(line_stream, item, ' ');
-    rig_from_world.rotation.w() = std::stold(item);
-
-    std::getline(line_stream, item, ' ');
-    rig_from_world.rotation.x() = std::stold(item);
-
-    std::getline(line_stream, item, ' ');
-    rig_from_world.rotation.y() = std::stold(item);
-
-    std::getline(line_stream, item, ' ');
-    rig_from_world.rotation.z() = std::stold(item);
-
-    std::getline(line_stream, item, ' ');
-    rig_from_world.translation.x() = std::stold(item);
-
-    std::getline(line_stream, item, ' ');
-    rig_from_world.translation.y() = std::stold(item);
-
-    std::getline(line_stream, item, ' ');
-    rig_from_world.translation.z() = std::stold(item);
-
-    frame.SetRigFromWorld(rig_from_world);
-
-    // DATA_IDS
-    std::getline(line_stream, item, ' ');
-    const uint32_t num_data_ids = std::stoul(item);
-    for (uint32_t i = 0; i < num_data_ids; ++i) {
-      data_t data_id;
-      std::getline(line_stream, item, ' ');
-      data_id.sensor_id.type = SensorTypeFromString(item);
-      std::getline(line_stream, item, ' ');
-      data_id.sensor_id.id = std::stoul(item);
-      std::getline(line_stream, item, ' ');
-      data_id.id = std::stoull(item);
-      frame.AddDataId(data_id);
-    }
-
-    reconstruction.AddFrame(std::move(frame));
-  }
-}
-
-void ReadFramesText(Reconstruction& reconstruction, const std::string& path) {
-  std::ifstream file(path);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadFramesText(reconstruction, file);
-}
-
-void ReadImagesText(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  // Handle backwards-compatibility for when we didn't have rigs and frames.
-  const bool is_legacy_reconstruction =
-      reconstruction.NumRigs() == 0 && reconstruction.NumFrames() == 0;
-  if (is_legacy_reconstruction) {
-    CreateOneRigPerCamera(reconstruction);
-  }
-
-  const std::unordered_map<image_t, Frame*> image_to_frame =
-      ExtractImageToFramePtr(reconstruction);
-
-  std::string line;
-  std::string item;
-
-  std::vector<Eigen::Vector2d> points2D;
-  std::vector<point3D_t> point3D_ids;
-
-  while (std::getline(stream, line)) {
-    StringTrim(&line);
-
-    if (line.empty() || line[0] == '#') {
-      continue;
-    }
-
-    std::stringstream line_stream1(line);
-
-    // ID
-    std::getline(line_stream1, item, ' ');
-    const image_t image_id = std::stoul(item);
-
-    class Image image;
-    image.SetImageId(image_id);
-
-    Rigid3d cam_from_world;
-
-    std::getline(line_stream1, item, ' ');
-    cam_from_world.rotation.w() = std::stold(item);
-
-    std::getline(line_stream1, item, ' ');
-    cam_from_world.rotation.x() = std::stold(item);
-
-    std::getline(line_stream1, item, ' ');
-    cam_from_world.rotation.y() = std::stold(item);
-
-    std::getline(line_stream1, item, ' ');
-    cam_from_world.rotation.z() = std::stold(item);
-
-    std::getline(line_stream1, item, ' ');
-    cam_from_world.translation.x() = std::stold(item);
-
-    std::getline(line_stream1, item, ' ');
-    cam_from_world.translation.y() = std::stold(item);
-
-    std::getline(line_stream1, item, ' ');
-    cam_from_world.translation.z() = std::stold(item);
-
-    // CAMERA_ID
-    std::getline(line_stream1, item, ' ');
-    image.SetCameraId(std::stoul(item));
-
-    if (is_legacy_reconstruction) {
-      CreateFrameForImage(image, cam_from_world, reconstruction);
-      image.SetFrameId(image.ImageId());
-      image.SetFramePtr(&reconstruction.Frame(image.ImageId()));
-    } else {
-      Frame* frame = image_to_frame.at(image.ImageId());
-      image.SetFrameId(frame->FrameId());
-      image.SetFramePtr(frame);
-    }
-
-    // NAME
-    std::getline(line_stream1, item, ' ');
-    image.SetName(item);
-
-    // POINTS2D
-    if (!std::getline(stream, line)) {
-      break;
-    }
-
-    StringTrim(&line);
-    std::stringstream line_stream2(line);
-
-    points2D.clear();
-    point3D_ids.clear();
-
-    if (!line.empty()) {
-      while (!line_stream2.eof()) {
-        Eigen::Vector2d point;
-
-        std::getline(line_stream2, item, ' ');
-        point.x() = std::stold(item);
-
-        std::getline(line_stream2, item, ' ');
-        point.y() = std::stold(item);
-
-        points2D.push_back(point);
-
-        std::getline(line_stream2, item, ' ');
-        if (item == "-1") {
-          point3D_ids.push_back(kInvalidPoint3DId);
-        } else {
-          point3D_ids.push_back(std::stoll(item));
-        }
-      }
-    }
-
-    image.SetPoints2D(points2D);
-
-    for (point2D_t point2D_idx = 0; point2D_idx < image.NumPoints2D();
-         ++point2D_idx) {
-      if (point3D_ids[point2D_idx] != kInvalidPoint3DId) {
-        image.SetPoint3DForPoint2D(point2D_idx, point3D_ids[point2D_idx]);
-      }
-    }
-
-    reconstruction.AddImage(std::move(image));
-  }
-}
-
-void ReadImagesText(Reconstruction& reconstruction, const std::string& path) {
-  std::ifstream file(path);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadImagesText(reconstruction, file);
-}
-
-void ReadPoints3DText(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  std::string line;
-  std::string item;
-
-  while (std::getline(stream, line)) {
-    StringTrim(&line);
-
-    if (line.empty() || line[0] == '#') {
-      continue;
-    }
-
-    std::stringstream line_stream(line);
-
-    // ID
-    std::getline(line_stream, item, ' ');
-    const point3D_t point3D_id = std::stoll(item);
-
-    struct Point3D point3D;
-
-    // XYZ
-    std::getline(line_stream, item, ' ');
-    point3D.xyz(0) = std::stold(item);
-
-    std::getline(line_stream, item, ' ');
-    point3D.xyz(1) = std::stold(item);
-
-    std::getline(line_stream, item, ' ');
-    point3D.xyz(2) = std::stold(item);
-
-    // Color
-    std::getline(line_stream, item, ' ');
-    point3D.color(0) = static_cast<uint8_t>(std::stoi(item));
-
-    std::getline(line_stream, item, ' ');
-    point3D.color(1) = static_cast<uint8_t>(std::stoi(item));
-
-    std::getline(line_stream, item, ' ');
-    point3D.color(2) = static_cast<uint8_t>(std::stoi(item));
-
-    // ERROR
-    std::getline(line_stream, item, ' ');
-    point3D.error = std::stold(item);
-
-    // TRACK
-    while (!line_stream.eof()) {
-      TrackElement track_el;
-
-      std::getline(line_stream, item, ' ');
-      StringTrim(&item);
-      if (item.empty()) {
-        break;
-      }
-      track_el.image_id = std::stoul(item);
-
-      std::getline(line_stream, item, ' ');
-      track_el.point2D_idx = std::stoul(item);
-
-      point3D.track.AddElement(track_el);
-    }
-
-    point3D.track.Compress();
-
-    reconstruction.AddPoint3D(point3D_id, std::move(point3D));
-  }
-}
-
-void ReadPoints3DText(Reconstruction& reconstruction, const std::string& path) {
-  std::ifstream file(path);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadPoints3DText(reconstruction, file);
-}
-
-void ReadRigsBinary(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  const size_t num_rigs = ReadBinaryLittleEndian<uint64_t>(&stream);
-  for (size_t i = 0; i < num_rigs; ++i) {
-    Rig rig;
-
-    rig.SetRigId(ReadBinaryLittleEndian<rig_t>(&stream));
-    const uint32_t num_sensors = ReadBinaryLittleEndian<uint32_t>(&stream);
-
-    if (num_sensors > 0) {
-      sensor_t ref_sensor_id;
-      ref_sensor_id.type =
-          static_cast<SensorType>(ReadBinaryLittleEndian<int>(&stream));
-      ref_sensor_id.id = ReadBinaryLittleEndian<uint32_t>(&stream);
-      rig.AddRefSensor(ref_sensor_id);
-    }
-
-    if (num_sensors > 1) {
-      for (uint32_t j = 0; j < num_sensors - 1; ++j) {
-        sensor_t sensor_id;
-        sensor_id.type =
-            static_cast<SensorType>(ReadBinaryLittleEndian<int>(&stream));
-        sensor_id.id = ReadBinaryLittleEndian<uint32_t>(&stream);
-
-        const bool has_pose = ReadBinaryLittleEndian<uint8_t>(&stream);
-        std::optional<Rigid3d> sensor_from_rig;
-        if (has_pose) {
-          sensor_from_rig = Rigid3d();
-          sensor_from_rig->rotation.w() =
-              ReadBinaryLittleEndian<double>(&stream);
-          sensor_from_rig->rotation.x() =
-              ReadBinaryLittleEndian<double>(&stream);
-          sensor_from_rig->rotation.y() =
-              ReadBinaryLittleEndian<double>(&stream);
-          sensor_from_rig->rotation.z() =
-              ReadBinaryLittleEndian<double>(&stream);
-          sensor_from_rig->translation.x() =
-              ReadBinaryLittleEndian<double>(&stream);
-          sensor_from_rig->translation.y() =
-              ReadBinaryLittleEndian<double>(&stream);
-          sensor_from_rig->translation.z() =
-              ReadBinaryLittleEndian<double>(&stream);
-        }
-
-        rig.AddSensor(sensor_id, sensor_from_rig);
-      }
-    }
-
-    reconstruction.AddRig(std::move(rig));
-  }
-}
-
-void ReadRigsBinary(Reconstruction& reconstruction, const std::string& path) {
-  std::ifstream file(path, std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadRigsBinary(reconstruction, file);
-}
-
-void ReadCamerasBinary(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  const size_t num_cameras = ReadBinaryLittleEndian<uint64_t>(&stream);
-  for (size_t i = 0; i < num_cameras; ++i) {
-    struct Camera camera;
-    camera.camera_id = ReadBinaryLittleEndian<camera_t>(&stream);
-    camera.model_id =
-        static_cast<CameraModelId>(ReadBinaryLittleEndian<int>(&stream));
-    camera.width = ReadBinaryLittleEndian<uint64_t>(&stream);
-    camera.height = ReadBinaryLittleEndian<uint64_t>(&stream);
-    camera.params.resize(CameraModelNumParams(camera.model_id), 0.);
-    ReadBinaryLittleEndian<double>(&stream, &camera.params);
-    THROW_CHECK(camera.VerifyParams());
-    reconstruction.AddCamera(std::move(camera));
-  }
-}
-
-void ReadCamerasBinary(Reconstruction& reconstruction,
-                       const std::string& path) {
-  std::ifstream file(path, std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadCamerasBinary(reconstruction, file);
-}
-
-void ReadFramesBinary(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  const uint64_t num_frames = ReadBinaryLittleEndian<uint64_t>(&stream);
-  for (uint64_t i = 0; i < num_frames; ++i) {
-    Frame frame;
-
-    frame.SetFrameId(ReadBinaryLittleEndian<frame_t>(&stream));
-    frame.SetRigId(ReadBinaryLittleEndian<rig_t>(&stream));
-
-    Rigid3d rig_from_world;
-    rig_from_world.rotation.w() = ReadBinaryLittleEndian<double>(&stream);
-    rig_from_world.rotation.x() = ReadBinaryLittleEndian<double>(&stream);
-    rig_from_world.rotation.y() = ReadBinaryLittleEndian<double>(&stream);
-    rig_from_world.rotation.z() = ReadBinaryLittleEndian<double>(&stream);
-    rig_from_world.translation.x() = ReadBinaryLittleEndian<double>(&stream);
-    rig_from_world.translation.y() = ReadBinaryLittleEndian<double>(&stream);
-    rig_from_world.translation.z() = ReadBinaryLittleEndian<double>(&stream);
-    frame.SetRigFromWorld(rig_from_world);
-
-    const uint32_t num_data_ids = ReadBinaryLittleEndian<uint32_t>(&stream);
-    for (uint32_t j = 0; j < num_data_ids; ++j) {
-      data_t data_id;
-      data_id.sensor_id.type =
-          static_cast<SensorType>(ReadBinaryLittleEndian<int>(&stream));
-      data_id.sensor_id.id = ReadBinaryLittleEndian<uint32_t>(&stream);
-      data_id.id = ReadBinaryLittleEndian<uint64_t>(&stream);
-      frame.AddDataId(data_id);
-    }
-
-    reconstruction.AddFrame(std::move(frame));
-  }
-}
-
-void ReadFramesBinary(Reconstruction& reconstruction, const std::string& path) {
-  std::ifstream file(path, std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadFramesBinary(reconstruction, file);
-}
-
-void ReadImagesBinary(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  // Handle backwards-compatibility for when we didn't have rigs and frames.
-  const bool is_legacy_reconstruction =
-      reconstruction.NumRigs() == 0 && reconstruction.NumFrames() == 0;
-  if (is_legacy_reconstruction) {
-    CreateOneRigPerCamera(reconstruction);
-  }
-
-  const std::unordered_map<image_t, Frame*> image_to_frame =
-      ExtractImageToFramePtr(reconstruction);
-
-  std::vector<Eigen::Vector2d> points2D;
-  std::vector<point3D_t> point3D_ids;
-
-  const size_t num_reg_images = ReadBinaryLittleEndian<uint64_t>(&stream);
-  for (size_t i = 0; i < num_reg_images; ++i) {
-    class Image image;
-
-    image.SetImageId(ReadBinaryLittleEndian<image_t>(&stream));
-
-    Rigid3d cam_from_world;
-    cam_from_world.rotation.w() = ReadBinaryLittleEndian<double>(&stream);
-    cam_from_world.rotation.x() = ReadBinaryLittleEndian<double>(&stream);
-    cam_from_world.rotation.y() = ReadBinaryLittleEndian<double>(&stream);
-    cam_from_world.rotation.z() = ReadBinaryLittleEndian<double>(&stream);
-    cam_from_world.translation.x() = ReadBinaryLittleEndian<double>(&stream);
-    cam_from_world.translation.y() = ReadBinaryLittleEndian<double>(&stream);
-    cam_from_world.translation.z() = ReadBinaryLittleEndian<double>(&stream);
-
-    image.SetCameraId(ReadBinaryLittleEndian<camera_t>(&stream));
-
-    if (is_legacy_reconstruction) {
-      CreateFrameForImage(image, cam_from_world, reconstruction);
-      image.SetFrameId(image.ImageId());
-      image.SetFramePtr(&reconstruction.Frame(image.ImageId()));
-    } else {
-      Frame* frame = image_to_frame.at(image.ImageId());
-      image.SetFrameId(frame->FrameId());
-      image.SetFramePtr(frame);
-    }
-
-    char name_char;
-    do {
-      stream.read(&name_char, 1);
-      if (name_char != '\0') {
-        image.Name() += name_char;
-      }
-    } while (name_char != '\0');
-
-    const size_t num_points2D = ReadBinaryLittleEndian<uint64_t>(&stream);
-
-    points2D.clear();
-    points2D.reserve(num_points2D);
-    point3D_ids.clear();
-    point3D_ids.reserve(num_points2D);
-    for (size_t j = 0; j < num_points2D; ++j) {
-      const double x = ReadBinaryLittleEndian<double>(&stream);
-      const double y = ReadBinaryLittleEndian<double>(&stream);
-      points2D.emplace_back(x, y);
-      point3D_ids.push_back(ReadBinaryLittleEndian<point3D_t>(&stream));
-    }
-
-    image.SetPoints2D(points2D);
-
-    for (point2D_t point2D_idx = 0; point2D_idx < image.NumPoints2D();
-         ++point2D_idx) {
-      if (point3D_ids[point2D_idx] != kInvalidPoint3DId) {
-        image.SetPoint3DForPoint2D(point2D_idx, point3D_ids[point2D_idx]);
-      }
-    }
-
-    reconstruction.AddImage(std::move(image));
-  }
-}
-
-void ReadImagesBinary(Reconstruction& reconstruction, const std::string& path) {
-  std::ifstream file(path, std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadImagesBinary(reconstruction, file);
-}
-
-void ReadPoints3DBinary(Reconstruction& reconstruction, std::istream& stream) {
-  THROW_CHECK(stream.good());
-
-  const size_t num_points3D = ReadBinaryLittleEndian<uint64_t>(&stream);
-  for (size_t i = 0; i < num_points3D; ++i) {
-    struct Point3D point3D;
-
-    const point3D_t point3D_id = ReadBinaryLittleEndian<point3D_t>(&stream);
-
-    point3D.xyz(0) = ReadBinaryLittleEndian<double>(&stream);
-    point3D.xyz(1) = ReadBinaryLittleEndian<double>(&stream);
-    point3D.xyz(2) = ReadBinaryLittleEndian<double>(&stream);
-    point3D.color(0) = ReadBinaryLittleEndian<uint8_t>(&stream);
-    point3D.color(1) = ReadBinaryLittleEndian<uint8_t>(&stream);
-    point3D.color(2) = ReadBinaryLittleEndian<uint8_t>(&stream);
-    point3D.error = ReadBinaryLittleEndian<double>(&stream);
-
-    const size_t track_length = ReadBinaryLittleEndian<uint64_t>(&stream);
-    for (size_t j = 0; j < track_length; ++j) {
-      const image_t image_id = ReadBinaryLittleEndian<image_t>(&stream);
-      const point2D_t point2D_idx = ReadBinaryLittleEndian<point2D_t>(&stream);
-      point3D.track.AddElement(image_id, point2D_idx);
-    }
-    point3D.track.Compress();
-
-    reconstruction.AddPoint3D(point3D_id, std::move(point3D));
-  }
-}
-
-void ReadPoints3DBinary(Reconstruction& reconstruction,
-                        const std::string& path) {
-  std::ifstream file(path, std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  ReadPoints3DBinary(reconstruction, file);
-}
-
-void WriteRigsText(const Reconstruction& reconstruction, std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  // Ensure that we don't loose any precision by storing in text.
-  stream.precision(17);
-
-  stream << "# Rig calib list with one line of data per calib:" << std::endl;
-  stream << "#   RIG_ID, NUM_SENSORS, REF_SENSOR_TYPE, REF_SENSOR_ID, "
-            "SENSORS[] as (SENSOR_TYPE, SENSOR_ID, HAS_POSE, [QW, QX, QY, QZ, "
-            "TX, TY, TZ])"
-         << std::endl;
-  stream << "# Number of rigs: " << reconstruction.NumRigs() << std::endl;
-
-  for (const camera_t rig_id : ExtractSortedIds(reconstruction.Rigs())) {
-    const Rig& rig = reconstruction.Rig(rig_id);
-
-    std::ostringstream line;
-    line.precision(17);
-
-    line << rig_id << " ";
-
-    line << rig.NumSensors() << " ";
-
-    if (rig.NumSensors() > 0) {
-      const sensor_t ref_sensor_id = rig.RefSensorId();
-      line << ref_sensor_id.type << " ";
-      line << ref_sensor_id.id << " ";
-    }
-
-    for (const auto& [sensor_id, sensor_from_rig] : rig.Sensors()) {
-      line << sensor_id.type << " ";
-      line << sensor_id.id << " ";
-      if (sensor_from_rig.has_value()) {
-        line << "1 ";
-        line << sensor_from_rig->rotation.w() << " ";
-        line << sensor_from_rig->rotation.x() << " ";
-        line << sensor_from_rig->rotation.y() << " ";
-        line << sensor_from_rig->rotation.z() << " ";
-        line << sensor_from_rig->translation.x() << " ";
-        line << sensor_from_rig->translation.y() << " ";
-        line << sensor_from_rig->translation.z() << " ";
-      } else {
-        line << "0 ";
-      }
-    }
-
-    std::string line_string = line.str();
-    line_string = line_string.substr(0, line_string.size() - 1);
-
-    stream << line_string << std::endl;
-  }
-}
-
-void WriteRigsText(const Reconstruction& reconstruction,
-                   const std::string& path) {
-  std::ofstream file(path, std::ios::trunc);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WriteRigsText(reconstruction, file);
-}
-
-void WriteCamerasText(const Reconstruction& reconstruction,
-                      std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  // Ensure that we don't loose any precision by storing in text.
-  stream.precision(17);
-
-  stream << "# Camera list with one line of data per camera:" << std::endl;
-  stream << "#   CAMERA_ID, MODEL, WIDTH, HEIGHT, PARAMS[]" << std::endl;
-  stream << "# Number of cameras: " << reconstruction.NumCameras() << std::endl;
-
-  for (const camera_t camera_id : ExtractSortedIds(reconstruction.Cameras())) {
-    const Camera& camera = reconstruction.Camera(camera_id);
-
-    std::ostringstream line;
-    line.precision(17);
-
-    line << camera_id << " ";
-    line << camera.ModelName() << " ";
-    line << camera.width << " ";
-    line << camera.height << " ";
-
-    for (const double param : camera.params) {
-      line << param << " ";
-    }
-
-    std::string line_string = line.str();
-    line_string = line_string.substr(0, line_string.size() - 1);
-
-    stream << line_string << std::endl;
-  }
-}
-
-void WriteCamerasText(const Reconstruction& reconstruction,
-                      const std::string& path) {
-  std::ofstream file(path, std::ios::trunc);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WriteCamerasText(reconstruction, file);
-}
-
-void WriteFramesText(const Reconstruction& reconstruction,
-                     std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  const std::vector<frame_t> frame_ids = ExtractSortedIds<frame_t, Frame>(
-      reconstruction.Frames(),
-      [](const Frame& frame) { return frame.HasPose(); });
-
-  // Ensure that we don't loose any precision by storing in text.
-  stream.precision(17);
-
-  stream << "# Frame list with one line of data per frame:" << std::endl;
-  stream << "#   FRAME_ID, RIG_ID, "
-            "RIG_FROM_WORLD[QW, QX, QY, QZ, TX, TY, TZ], NUM_DATA_IDS, "
-            "DATA_IDS[] as (SENSOR_TYPE, SENSOR_ID, DATA_ID)"
-         << std::endl;
-  stream << "# Number of frames: " << frame_ids.size() << std::endl;
-
-  stream.precision(17);
-
-  for (const frame_t frame_id : frame_ids) {
-    const Frame& frame = reconstruction.Frame(frame_id);
-
-    stream << frame_id << " ";
-    stream << frame.RigId() << " ";
-
-    const Rigid3d& rig_from_world = frame.RigFromWorld();
-    stream << rig_from_world.rotation.w() << " ";
-    stream << rig_from_world.rotation.x() << " ";
-    stream << rig_from_world.rotation.y() << " ";
-    stream << rig_from_world.rotation.z() << " ";
-    stream << rig_from_world.translation.x() << " ";
-    stream << rig_from_world.translation.y() << " ";
-    stream << rig_from_world.translation.z() << " ";
-
-    const std::set<data_t>& data_ids = frame.DataIds();
-    stream << data_ids.size();
-    for (const data_t& data_id : data_ids) {
-      stream << " " << data_id.sensor_id.type << " " << data_id.sensor_id.id
-             << " " << data_id.id;
-    }
-
-    stream << std::endl;
-  }
-}
-
-void WriteFramesText(const Reconstruction& reconstruction,
-                     const std::string& path) {
-  std::ofstream file(path, std::ios::trunc);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WriteFramesText(reconstruction, file);
-}
-
-void WriteImagesText(const Reconstruction& reconstruction,
-                     std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  // Ensure that we don't loose any precision by storing in text.
-  stream.precision(17);
-
-  stream << "# Image list with two lines of data per image:" << std::endl;
-  stream << "#   IMAGE_ID, QW, QX, QY, QZ, TX, TY, TZ, CAMERA_ID, "
-            "NAME"
-         << std::endl;
-  stream << "#   POINTS2D[] as (X, Y, POINT3D_ID)" << std::endl;
-  stream << "# Number of images: " << reconstruction.NumRegImages()
-         << ", mean observations per image: "
-         << reconstruction.ComputeMeanObservationsPerRegImage() << std::endl;
-
-  std::ostringstream line;
-  line.precision(17);
-
-  for (const image_t image_id : reconstruction.RegImageIds()) {
-    const Image& image = reconstruction.Image(image_id);
-
-    line.str("");
-    line.clear();
-
-    line << image_id << " ";
-
-    const Rigid3d& cam_from_world = image.CamFromWorld();
-    line << cam_from_world.rotation.w() << " ";
-    line << cam_from_world.rotation.x() << " ";
-    line << cam_from_world.rotation.y() << " ";
-    line << cam_from_world.rotation.z() << " ";
-    line << cam_from_world.translation.x() << " ";
-    line << cam_from_world.translation.y() << " ";
-    line << cam_from_world.translation.z() << " ";
-
-    line << image.CameraId() << " ";
-
-    line << image.Name();
-
-    stream << line.str() << std::endl;
-
-    line.str("");
-    line.clear();
-
-    for (const Point2D& point2D : image.Points2D()) {
-      line << point2D.xy(0) << " ";
-      line << point2D.xy(1) << " ";
-      if (point2D.HasPoint3D()) {
-        line << point2D.point3D_id << " ";
-      } else {
-        line << -1 << " ";
-      }
-    }
-    if (image.NumPoints2D() > 0) {
-      line.seekp(-1, std::ios_base::end);
-    }
-    stream << line.str() << std::endl;
-  }
-}
-
-void WriteImagesText(const Reconstruction& reconstruction,
-                     const std::string& path) {
-  std::ofstream file(path, std::ios::trunc);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WriteImagesText(reconstruction, file);
-}
-
-void WritePoints3DText(const Reconstruction& reconstruction,
-                       std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  // Ensure that we don't loose any precision by storing in text.
-  stream.precision(17);
-
-  stream << "# 3D point list with one line of data per point:" << std::endl;
-  stream << "#   POINT3D_ID, X, Y, Z, R, G, B, ERROR, "
-            "TRACK[] as (IMAGE_ID, POINT2D_IDX)"
-         << std::endl;
-  stream << "# Number of points: " << reconstruction.NumPoints3D()
-         << ", mean track length: " << reconstruction.ComputeMeanTrackLength()
-         << std::endl;
-
-  for (const point3D_t point3D_id :
-       ExtractSortedIds(reconstruction.Points3D())) {
-    const Point3D& point3D = reconstruction.Point3D(point3D_id);
-
-    stream << point3D_id << " ";
-    stream << point3D.xyz(0) << " ";
-    stream << point3D.xyz(1) << " ";
-    stream << point3D.xyz(2) << " ";
-    stream << static_cast<int>(point3D.color(0)) << " ";
-    stream << static_cast<int>(point3D.color(1)) << " ";
-    stream << static_cast<int>(point3D.color(2)) << " ";
-    stream << point3D.error << " ";
-
-    std::ostringstream line;
-    line.precision(17);
-
-    for (const auto& track_el : point3D.track.Elements()) {
-      line << track_el.image_id << " ";
-      line << track_el.point2D_idx << " ";
-    }
-
-    std::string line_string = line.str();
-    line_string = line_string.substr(0, line_string.size() - 1);
-
-    stream << line_string << std::endl;
-  }
-}
-
-void WritePoints3DText(const Reconstruction& reconstruction,
-                       const std::string& path) {
-  std::ofstream file(path, std::ios::trunc);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WritePoints3DText(reconstruction, file);
-}
-
-void WriteRigsBinary(const Reconstruction& reconstruction,
-                     std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  WriteBinaryLittleEndian<uint64_t>(&stream, reconstruction.NumRigs());
-
-  for (const rig_t rig_id : ExtractSortedIds(reconstruction.Rigs())) {
-    const Rig& rig = reconstruction.Rig(rig_id);
-    WriteBinaryLittleEndian<rig_t>(&stream, rig_id);
-    WriteBinaryLittleEndian<uint32_t>(&stream, rig.NumSensors());
-    if (rig.NumSensors() > 0) {
-      WriteBinaryLittleEndian<int>(&stream,
-                                   static_cast<int>(rig.RefSensorId().type));
-      WriteBinaryLittleEndian<uint32_t>(&stream, rig.RefSensorId().id);
-    }
-    for (const auto& [sensor_id, sensor_from_rig] : rig.Sensors()) {
-      WriteBinaryLittleEndian<int>(&stream, static_cast<int>(sensor_id.type));
-      WriteBinaryLittleEndian<uint32_t>(&stream, sensor_id.id);
-      WriteBinaryLittleEndian<uint8_t>(&stream,
-                                       sensor_from_rig.has_value() ? 1 : 0);
-      if (sensor_from_rig.has_value()) {
-        WriteBinaryLittleEndian<double>(&stream, sensor_from_rig->rotation.w());
-        WriteBinaryLittleEndian<double>(&stream, sensor_from_rig->rotation.x());
-        WriteBinaryLittleEndian<double>(&stream, sensor_from_rig->rotation.y());
-        WriteBinaryLittleEndian<double>(&stream, sensor_from_rig->rotation.z());
-        WriteBinaryLittleEndian<double>(&stream,
-                                        sensor_from_rig->translation.x());
-        WriteBinaryLittleEndian<double>(&stream,
-                                        sensor_from_rig->translation.y());
-        WriteBinaryLittleEndian<double>(&stream,
-                                        sensor_from_rig->translation.z());
-      }
-    }
-  }
-}
-
-void WriteRigsBinary(const Reconstruction& reconstruction,
-                     const std::string& path) {
-  std::ofstream file(path, std::ios::trunc | std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WriteRigsBinary(reconstruction, file);
-}
-
-void WriteCamerasBinary(const Reconstruction& reconstruction,
-                        std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  WriteBinaryLittleEndian<uint64_t>(&stream, reconstruction.NumCameras());
-
-  for (const camera_t camera_id : ExtractSortedIds(reconstruction.Cameras())) {
-    const Camera& camera = reconstruction.Camera(camera_id);
-    WriteBinaryLittleEndian<camera_t>(&stream, camera_id);
-    WriteBinaryLittleEndian<int>(&stream, static_cast<int>(camera.model_id));
-    WriteBinaryLittleEndian<uint64_t>(&stream, camera.width);
-    WriteBinaryLittleEndian<uint64_t>(&stream, camera.height);
-    for (const double param : camera.params) {
-      WriteBinaryLittleEndian<double>(&stream, param);
-    }
-  }
-}
-
-void WriteCamerasBinary(const Reconstruction& reconstruction,
-                        const std::string& path) {
-  std::ofstream file(path, std::ios::trunc | std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WriteCamerasBinary(reconstruction, file);
-}
-
-void WriteFramesBinary(const Reconstruction& reconstruction,
-                       std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  const std::vector<frame_t> frame_ids = ExtractSortedIds<frame_t, Frame>(
-      reconstruction.Frames(),
-      [](const Frame& frame) { return frame.HasPose(); });
-
-  WriteBinaryLittleEndian<uint64_t>(&stream, frame_ids.size());
-
-  for (const frame_t frame_id : frame_ids) {
-    const Frame& frame = reconstruction.Frame(frame_id);
-
-    WriteBinaryLittleEndian<frame_t>(&stream, frame_id);
-    WriteBinaryLittleEndian<rig_t>(&stream, frame.RigId());
-
-    const Rigid3d& rig_from_world = frame.RigFromWorld();
-    WriteBinaryLittleEndian<double>(&stream, rig_from_world.rotation.w());
-    WriteBinaryLittleEndian<double>(&stream, rig_from_world.rotation.x());
-    WriteBinaryLittleEndian<double>(&stream, rig_from_world.rotation.y());
-    WriteBinaryLittleEndian<double>(&stream, rig_from_world.rotation.z());
-    WriteBinaryLittleEndian<double>(&stream, rig_from_world.translation.x());
-    WriteBinaryLittleEndian<double>(&stream, rig_from_world.translation.y());
-    WriteBinaryLittleEndian<double>(&stream, rig_from_world.translation.z());
-
-    const std::set<data_t>& data_ids = frame.DataIds();
-    WriteBinaryLittleEndian<uint32_t>(&stream, data_ids.size());
-    for (const data_t& data_id : data_ids) {
-      WriteBinaryLittleEndian<int>(&stream,
-                                   static_cast<int>(data_id.sensor_id.type));
-      WriteBinaryLittleEndian<uint32_t>(&stream, data_id.sensor_id.id);
-      WriteBinaryLittleEndian<uint64_t>(&stream, data_id.id);
-    }
-  }
-}
-
-void WriteFramesBinary(const Reconstruction& reconstruction,
-                       const std::string& path) {
-  std::ofstream file(path, std::ios::trunc | std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WriteFramesBinary(reconstruction, file);
-}
-
-void WriteImagesBinary(const Reconstruction& reconstruction,
-                       std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  WriteBinaryLittleEndian<uint64_t>(&stream, reconstruction.NumRegImages());
-
-  for (const image_t image_id : reconstruction.RegImageIds()) {
-    const Image& image = reconstruction.Image(image_id);
-
-    WriteBinaryLittleEndian<image_t>(&stream, image_id);
-
-    const Rigid3d& cam_from_world = image.CamFromWorld();
-    WriteBinaryLittleEndian<double>(&stream, cam_from_world.rotation.w());
-    WriteBinaryLittleEndian<double>(&stream, cam_from_world.rotation.x());
-    WriteBinaryLittleEndian<double>(&stream, cam_from_world.rotation.y());
-    WriteBinaryLittleEndian<double>(&stream, cam_from_world.rotation.z());
-    WriteBinaryLittleEndian<double>(&stream, cam_from_world.translation.x());
-    WriteBinaryLittleEndian<double>(&stream, cam_from_world.translation.y());
-    WriteBinaryLittleEndian<double>(&stream, cam_from_world.translation.z());
-
-    WriteBinaryLittleEndian<camera_t>(&stream, image.CameraId());
-
-    const std::string name = image.Name() + '\0';
-    stream.write(name.c_str(), name.size());
-
-    WriteBinaryLittleEndian<uint64_t>(&stream, image.NumPoints2D());
-    for (const Point2D& point2D : image.Points2D()) {
-      WriteBinaryLittleEndian<double>(&stream, point2D.xy(0));
-      WriteBinaryLittleEndian<double>(&stream, point2D.xy(1));
-      WriteBinaryLittleEndian<point3D_t>(&stream, point2D.point3D_id);
-    }
-  }
-}
-
-void WriteImagesBinary(const Reconstruction& reconstruction,
-                       const std::string& path) {
-  std::ofstream file(path, std::ios::trunc | std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WriteImagesBinary(reconstruction, file);
-}
-
-void WritePoints3DBinary(const Reconstruction& reconstruction,
-                         std::ostream& stream) {
-  THROW_CHECK(stream.good());
-
-  WriteBinaryLittleEndian<uint64_t>(&stream, reconstruction.NumPoints3D());
-
-  for (const point3D_t point3D_id :
-       ExtractSortedIds(reconstruction.Points3D())) {
-    const Point3D& point3D = reconstruction.Point3D(point3D_id);
-
-    WriteBinaryLittleEndian<point3D_t>(&stream, point3D_id);
-    WriteBinaryLittleEndian<double>(&stream, point3D.xyz(0));
-    WriteBinaryLittleEndian<double>(&stream, point3D.xyz(1));
-    WriteBinaryLittleEndian<double>(&stream, point3D.xyz(2));
-    WriteBinaryLittleEndian<uint8_t>(&stream, point3D.color(0));
-    WriteBinaryLittleEndian<uint8_t>(&stream, point3D.color(1));
-    WriteBinaryLittleEndian<uint8_t>(&stream, point3D.color(2));
-    WriteBinaryLittleEndian<double>(&stream, point3D.error);
-
-    WriteBinaryLittleEndian<uint64_t>(&stream, point3D.track.Length());
-    for (const TrackElement& track_el : point3D.track.Elements()) {
-      WriteBinaryLittleEndian<image_t>(&stream, track_el.image_id);
-      WriteBinaryLittleEndian<point2D_t>(&stream, track_el.point2D_idx);
-    }
-  }
-}
-
-void WritePoints3DBinary(const Reconstruction& reconstruction,
-                         const std::string& path) {
-  std::ofstream file(path, std::ios::trunc | std::ios::binary);
-  THROW_CHECK_FILE_OPEN(file, path);
-  WritePoints3DBinary(reconstruction, file);
-}
 
 bool ExportNVM(const Reconstruction& reconstruction,
                const std::string& path,
