+++ src/colmap/scene/reconstruction_io.h
@@ -30,94 +30,14 @@
 #pragma once
 
 #include "colmap/scene/reconstruction.h"
+#include "colmap/scene/reconstruction_io_binary.h"
+#include "colmap/scene/reconstruction_io_text.h"
 
 #include <iostream>
 
 #include <Eigen/Core>
 
 namespace colmap {
-
-// Note that cameras must be read before images.
-
-void ReadRigsText(Reconstruction& reconstruction, std::istream& stream);
-void ReadRigsText(Reconstruction& reconstruction, const std::string& path);
-
-void ReadCamerasText(Reconstruction& reconstruction, std::istream& stream);
-void ReadCamerasText(Reconstruction& reconstruction, const std::string& path);
-
-void ReadFramesText(Reconstruction& reconstruction, std::istream& stream);
-void ReadFramesText(Reconstruction& reconstruction, const std::string& path);
-
-void ReadImagesText(Reconstruction& reconstruction, std::istream& stream);
-void ReadImagesText(Reconstruction& reconstruction, const std::string& path);
-
-void ReadPoints3DText(Reconstruction& reconstruction, std::istream& stream);
-void ReadPoints3DText(Reconstruction& reconstruction, const std::string& path);
-
-void ReadRigsBinary(Reconstruction& reconstruction, std::istream& stream);
-void ReadRigsBinary(Reconstruction& reconstruction, const std::string& path);
-
-void ReadCamerasBinary(Reconstruction& reconstruction, std::istream& stream);
-void ReadCamerasBinary(Reconstruction& reconstruction, const std::string& path);
-
-void ReadFramesBinary(Reconstruction& reconstruction, std::istream& stream);
-void ReadFramesBinary(Reconstruction& reconstruction, const std::string& path);
-
-void ReadImagesBinary(Reconstruction& reconstruction, std::istream& stream);
-void ReadImagesBinary(Reconstruction& reconstruction, const std::string& path);
-
-void ReadPoints3DBinary(Reconstruction& reconstruction, std::istream& stream);
-void ReadPoints3DBinary(Reconstruction& reconstruction,
-                        const std::string& path);
-
-void WriteRigsText(const Reconstruction& reconstruction, std::ostream& stream);
-void WriteRigsText(const Reconstruction& reconstruction,
-                   const std::string& path);
-
-void WriteCamerasText(const Reconstruction& reconstruction,
-                      std::ostream& stream);
-void WriteCamerasText(const Reconstruction& reconstruction,
-                      const std::string& path);
-
-void WriteFramesText(const Reconstruction& reconstruction,
-                     std::ostream& stream);
-void WriteFramesText(const Reconstruction& reconstruction,
-                     const std::string& path);
-
-void WriteImagesText(const Reconstruction& reconstruction,
-                     std::ostream& stream);
-void WriteImagesText(const Reconstruction& reconstruction,
-                     const std::string& path);
-
-void WritePoints3DText(const Reconstruction& reconstruction,
-                       std::ostream& stream);
-void WritePoints3DText(const Reconstruction& reconstruction,
-                       const std::string& path);
-
-void WriteRigsBinary(const Reconstruction& reconstruction,
-                     std::ostream& stream);
-void WriteRigsBinary(const Reconstruction& reconstruction,
-                     const std::string& path);
-
-void WriteFramesBinary(const Reconstruction& reconstruction,
-                       std::ostream& stream);
-void WriteFramesBinary(const Reconstruction& reconstruction,
-                       const std::string& path);
-
-void WriteCamerasBinary(const Reconstruction& reconstruction,
-                        std::ostream& stream);
-void WriteCamerasBinary(const Reconstruction& reconstruction,
-                        const std::string& path);
-
-void WriteImagesBinary(const Reconstruction& reconstruction,
-                       std::ostream& stream);
-void WriteImagesBinary(const Reconstruction& reconstruction,
-                       const std::string& path);
-
-void WritePoints3DBinary(const Reconstruction& reconstruction,
-                         std::ostream& stream);
-void WritePoints3DBinary(const Reconstruction& reconstruction,
-                         const std::string& path);
 
 // Exports in NVM format http://ccwu.me/vsfm/doc.html#nvm. Only supports
 // SIMPLE_RADIAL camera model when exporting distortion parameters. When
